chore(SG-236): фdd additional metricssqlrepository, write to both in the service
