+++ src/app/providers.ts
@@ -16,10 +16,11 @@ import { MetricTypeService } from "../modules/metric-type/metric-type.service";
 import { MetricTypeRepository } from "../modules/metric-type/metric-type.repository";
 import { MetricCategoryService } from "../modules/metric-category/metric-category.service";
 import { MetricCategoryRepository } from "../modules/metric-category/metric-category.repository";
-import { MetricTimestreamRepository } from "../modules/metric/repositories/metric-timestream.repository";
+// import { MetricTimestreamRepository } from "../modules/metric/repositories/metric-timestream.repository";
+import { MetricRepository } from "../modules/metric/metric.repository";
 import { ReportsRepository } from "../modules/reports/reports.repository";
 import { ReportsService } from "../modules/reports/reports.service";
-import { MetricAuroraRepository } from "../modules/metric/repositories/metric-aurora.repository";
+// import { MetricSqlRepository } from "../modules/metric/repositories/metric-sql.repository";
 
 // load and validate env variables
 loadEnvironment();
@@ -63,6 +64,7 @@ container.register("EventbusService", EventbusService);
 // repositories
 container.register("MetricCategoryRepository", MetricCategoryRepository);
 container.register("MetricTypeRepository", MetricTypeRepository);
-container.register("MetricRepository", MetricTimestreamRepository);
+container.register("MetricRepository", MetricRepository);
+// container.register("MetricTimestreamRepository", MetricTimestreamRepository);
 container.register("ReportsRepository", ReportsRepository);
-container.register("MetricAuroraRepository", MetricAuroraRepository);
+// container.register("MetricAuroraRepository", MetricSqlRepository);
